Move make_qr_image to stylers.py
